build: update ts defs and docs parser (#43985)
* build: update ts defs and docs parser

* build: update dig
