+++ .github/workflows/archaeologist-dig.yml
@@ -12,6 +12,10 @@ jobs:
         uses: actions/checkout@692973e3d937129bcbf40652eb9f2f61becf3332 #v4.0.2
         with:
           fetch-depth: 0
+      - name: Setup Node.js/npm
+        uses: actions/setup-node@0a44ba7841725637a19e28fa30b79a866c81b0a6
+        with:
+          node-version: 20.11.x
       - name: Setting Up Dig Site
         run: |
           echo "remote: ${{ github.event.pull_request.head.repo.clone_url }}"
