+++ package.json
@@ -6,11 +6,11 @@
   "devDependencies": {
     "@azure/storage-blob": "^12.25.0",
     "@electron/asar": "^3.2.13",
-    "@electron/docs-parser": "^1.2.3",
+    "@electron/docs-parser": "^2.0.0",
     "@electron/fiddle-core": "^1.3.4",
     "@electron/github-app-auth": "^2.2.1",
     "@electron/lint-roller": "^2.4.0",
-    "@electron/typescript-definitions": "^8.15.8",
+    "@electron/typescript-definitions": "^9.0.0",
     "@octokit/rest": "^20.0.2",
     "@primer/octicons": "^10.0.0",
     "@types/minimist": "^1.2.5",
@@ -81,7 +81,7 @@
     "lint:ts-check-js-in-markdown": "lint-roller-markdown-ts-check --root docs \"**/*.md\" --ignore \"breaking-changes.md\"",
     "lint:js-in-markdown": "lint-roller-markdown-standard --root docs \"**/*.md\"",
     "lint:api-history": "lint-roller-markdown-api-history --root \"./docs/api/\" --schema \"./docs/api-history.schema.json\" --breaking-changes-file \"./docs/breaking-changes.md\" --check-placement --check-strings \"*.md\"",
-    "create-api-json": "node script/create-api-json.js",
+    "create-api-json": "node script/create-api-json.mjs",
     "create-typescript-definitions": "npm run create-api-json && electron-typescript-definitions --api=electron-api.json && node spec/ts-smoke/runner.js",
     "gn-typescript-definitions": "npm run create-typescript-definitions && shx cp electron.d.ts",
     "pre-flight": "pre-flight",
